requirements.txt
Flask
matplotlib
nba_api
python-dotenv
plotly